+++ lib/shrine/plugins/sequel.rb
@@ -113,8 +113,6 @@ class Shrine
       end
 
       module AttacherMethods
-        private
-
         # Proceeds with updating the record unless the attachment has changed.
         def swap(uploaded_file)
           return if record.send(:"#{name}_data") != record.reload.send(:"#{name}_data")
@@ -123,6 +121,8 @@ class Shrine
         rescue ::Sequel::Error => error
           raise unless error.message == "Record not found" # prior to version 4.28
         end
+
+        private
 
         # Saves the record after assignment, skipping validations.
         def update(uploaded_file)