+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## HEAD
 
+* Set default UNIX permissions to 0644 for files and 0755 for directories (janko-m)
+
 * Apply directory permissions to all subfolders inside the main folder (janko-m)
 
 * Add `:directory_permissions` to `Storage::FileSystem` (janko-m)