+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## HEAD
+## 2.3.0 (2016-08-27)
 
 * Prevent client from caching the presign response in direct_upload plugin (janko-m)